refactor: extract Chief-of-Staff request models to a sibling module
Move ChatRequest and ConversationalProposeRequest out of the meta controller into _meta_requests.py so the controller stays under its module-size budget, and annotate the test's set_many override.
