+++ src/synthorg/api/controllers/meta.py
@@ -1,11 +1,9 @@
 """Meta improvement controller -- self-improvement proposals and signals."""
 
 from typing import Final
-from uuid import UUID
 
 from litestar import Controller, get, post
 from litestar.datastructures import State
-from pydantic import BaseModel, ConfigDict, Field
 
 from synthorg._core.features import require_service
 from synthorg.api._feature_gate import ensure_feature_enabled
@@ -16,6 +14,10 @@ from synthorg.api.controllers._meta_chat_window import resolve_chat_snapshot_win
 from synthorg.api.controllers._meta_proposal_helpers import (
     PROPOSAL_ACTION_TYPES,
     proposal_to_dict,
+)
+from synthorg.api.controllers._meta_requests import (
+    ChatRequest,
+    ConversationalProposeRequest,
 )
 from synthorg.api.cursor import decode_cursor
 from synthorg.api.dto import ApiResponse, PaginatedResponse
@@ -53,45 +55,6 @@ from synthorg.observability.events.meta import (
     META_PROPOSAL_LISTED,
 )
 from synthorg.persistence.state import persistence_of
-
-
-class ChatRequest(BaseModel):
-    """Request body for the Chief of Staff chat endpoint."""
-
-    model_config = ConfigDict(frozen=True, allow_inf_nan=False, extra="forbid")
-
-    question: NotBlankStr = Field(
-        max_length=2000,
-        description="Free-text question for the Chief of Staff agent.",
-    )
-    proposal_id: UUID | None = Field(
-        default=None,
-        description="Improvement proposal the question is scoped to, if any.",
-    )
-    alert_id: UUID | None = Field(
-        default=None,
-        description="Proactive alert the question is scoped to, if any.",
-    )
-
-
-class ConversationalProposeRequest(BaseModel):
-    """Request body for the Chief of Staff clarify-and-propose endpoint."""
-
-    model_config = ConfigDict(frozen=True, allow_inf_nan=False, extra="forbid")
-
-    message: NotBlankStr = Field(
-        max_length=2000,
-        description="Human message for the clarify-and-propose turn.",
-    )
-    conversation_id: NotBlankStr | None = Field(
-        default=None,
-        description="Existing conversation to append to; None starts a new one.",
-    )
-    project: NotBlankStr | None = Field(
-        default=None,
-        description="Project the proposal should be scoped to, if any.",
-    )
-
 
 logger = get_logger(__name__)
